requirements.txt
aio-pika==9.5.7
aiormq==6.9.0
async-timeout==5.0.1
asyncpg==0.30.0
exceptiongroup==1.3.0
greenlet==3.2.4
idna==3.11
multidict==6.7.0
pamqp==3.3.0
propcache==0.4.1
python-dotenv==1.1.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
yarl==1.22.0
